reused the textwidget
