+++ home_insurance_app/lib/pages/common_widgets.dart
@@ -21,3 +21,57 @@ class _CommonAppBarState extends State<CommonAppBar> {
     );
   }
 }
+
+class TextWidget extends StatelessWidget {
+  TextWidget(
+      {Key key,
+      @required String leftText,
+      @required String rightText,
+      Color leftColor,
+      Color rightColor})
+      : _leftText = leftText,
+        _rightText = rightText,
+        _leftColor = leftColor,
+        _rightColor = rightColor,
+        super(key: key);
+
+  final String _leftText;
+  final String _rightText;
+  final Color _leftColor;
+  final Color _rightColor;
+  @override
+  Widget build(BuildContext context) {
+    final double _padding = 18.0;
+
+    return Container(
+      child: Card(
+        elevation: 2.0,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
+        child: Padding(
+          padding: EdgeInsets.all(_padding),
+          child: Wrap(
+            alignment: WrapAlignment.spaceBetween,
+            // mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Text(
+                _leftText,
+                style: TextStyle(
+                  fontSize: 18.0,
+                  fontWeight: FontWeight.bold,
+                  color: _leftColor ?? Colors.grey[800],
+                ),
+              ),
+              Text(
+                _rightText,
+                style: TextStyle(
+                    fontSize: 20.0,
+                    fontWeight: FontWeight.w500,
+                    color: _rightColor ?? Colors.brown[600]),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
